make_collectionview_look_better Now the images are contained inside the cells and don't flow outside anymore.
The image view is now added to the contentView of the cell, instead of directly to the cell.
The cell is now set to clip to bounds, so it will not show any content outside of its bounds.
Next I'll have to constrain the image view to the size of the cell.