make_collectionview_look_better: Made the cells in the collection view larger; only 3 images are not fitting on a row. Also made the inset of the collection view equal to 5. Also set to 5 the spacing between cells and the spacing between lines so, now all the space arround cells is the same.
Made the collection view of the same size as the pullUpView as it was larger and you were not able to scroll and see all the retrieved images.

Lost the ability to close the collection by pulling down gesture; this could be because now the collection view has to be scrolable and it's the same gesture. We'll have to see how can we fix that.